refactor(upload-tool): 拆分component.ts
