+++ src/gadgets/upload-tool/i18n.ts
@@ -78,8 +78,8 @@ const MESSAGES = {
 	'btn-reset-preview': tr('重置为自动生成', '重設為自動產生'),
 	'ai-generated-label': '由AI生成',
 	'uploadtext-title': tr('上传说明', '上傳說明'),
-	'btn-back-to-native': tr('返回旧版表单', '返回舊版表單'),
-	'btn-batch-upload': tr('批量上传文件', '批次上傳檔案', '批次上載'),
+	'btn-back-to-native': tr('返回旧版', '返回舊版'),
+	'btn-batch-upload': tr('批量上传', '批次上傳', '批次上載'),
 	'note-section': tr('备注', '備註'),
 	'placeholder-note': tr('编辑摘要', '編輯摘要'),
 	'options-section': tr('上传选项', '上傳選項'),
